Don't include goffice's format-impl.h, (format_value_common): New
2006-12-18  Morten Welinder  <terra@gnome.org>

	* src/gnm-format.c: Don't include goffice's format-impl.h,
	(format_value_common): New function, mostly
	extracted from gnm_format_layout, but using
	go_format_value_gstring.
	(find_entry, gnm_style_format_condition): Remove.
	(hash_fill): Generalize to handle layouts and strings alike.
	(format_value_gstring): Simplify using format_value_common.

	* src/workbook-edit.c: Don't include goffice's format-impl.h,

+++ src/gnm-format.h
@@ -7,13 +7,14 @@
 
 char  *format_value	    (GOFormat const *format,
 			     GnmValue const *value, GOColor *go_color,
-			     double col_width, GODateConventions const *date_conv);
+			     int col_width,
+			     GODateConventions const *date_conv);
 
 GOFormatNumberError format_value_gstring (GString *result,
 					  GOFormat const *format,
 					  GnmValue const *value,
 					  GOColor *go_color,
-					  double col_width,
+					  int col_width,
 					  GODateConventions const *date_conv);
 
 GOFormatNumberError gnm_format_layout    (PangoLayout *result,
